exception handling eklendi, resource filter düzenlendi. Bu geliştirmelere bağlı olarak package yapısı değiştirildi.
