+++ microservices-demo-app/src/main/java/tr/com/innova/lega/demo/api/kisi/KisiDTO.java
@@ -1,4 +1,4 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.kisi;
 
 import com.fasterxml.jackson.annotation.JsonFilter;
 import lombok.Getter;
@@ -8,6 +8,7 @@ import lombok.Setter;
 @Setter
 @JsonFilter("KisiDTOFilter")
 public class KisiDTO {
+    public static final String FILTER_NAME = "KisiDTOFilter";
     private String id;
     private String ad;
     private String soyad;
