+++ microservices-demo-app/src/main/java/tr/com/innova/lega/demo/exception/ExceptionResponse.java
@@ -1,0 +1,21 @@
+package tr.com.innova.lega.demo.exception;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+public class ExceptionResponse {
+    private Date timestamp;
+    private String message;
+    private String details;
+
+    public ExceptionResponse(Date timestamp, String message, String details) {
+        super();
+        this.timestamp = timestamp;
+        this.message = message;
+        this.details = details;
+    }
+}
