+++ microservices-demo-app/src/main/resources/application.properties
@@ -10,4 +10,3 @@ eureka.client.register-with-eureka=true
 eureka.client.fetch-registry=true
 #register url to server
 eureka.client.serviceUrl.defaultZone=http://localhost:8080/eureka/
-instance.hostname=localhost
