+++ microservices-demo-app/src/test/java/tr/com/innova/lega/demo/api/kisi/KisiDTOTest.java
@@ -1,4 +1,4 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.kisi;
 
 import org.junit.Before;
 import org.junit.Test;
