+++ microservices-demo-client/src/main/java/tr/com/innova/lega/demo/api/adres/AdresRepository.java
@@ -1,8 +1,8 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.adres;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface KisiRepository extends JpaRepository<Kisi, String> {
+public interface AdresRepository extends JpaRepository<Adres, String> {
 }
