+++ microservices-demo-client/src/main/java/tr/com/innova/lega/demo/api/kisi/Kisi.java
@@ -1,4 +1,4 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.kisi;
 
 import lombok.Getter;
 import lombok.Setter;
