+++ microservices-demo-client/src/main/java/tr/com/innova/lega/demo/api/kisi/KisiRepository.java
@@ -1,8 +1,8 @@
-package tr.com.innova.lega.demo.adres;
+package tr.com.innova.lega.demo.api.kisi;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface AdresRepository extends JpaRepository<Adres, String> {
+public interface KisiRepository extends JpaRepository<Kisi, String> {
 }
