+++ microservices-demo-client/src/main/java/tr/com/innova/lega/demo/api/kisi/KisiService.java
@@ -1,4 +1,4 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.kisi;
 
 import tr.com.innova.lega.demo.base.AbstractService;
