+++ microservices-demo-service/src/main/java/tr/com/innova/lega/demo/api/adres/AdresServiceImpl.java
@@ -1,4 +1,4 @@
-package tr.com.innova.lega.demo.kisi;
+package tr.com.innova.lega.demo.api.adres;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -7,10 +7,10 @@ import tr.com.innova.lega.demo.base.AbstractServiceImpl;
 
 @Service
 @Transactional
-public class KisiServiceImpl extends AbstractServiceImpl<Kisi> implements KisiService {
+public class AdresServiceImpl extends AbstractServiceImpl<Adres> implements AdresService {
 
     @Autowired
-    public KisiServiceImpl(KisiRepository kisiRepository) {
-        super(kisiRepository);
+    public AdresServiceImpl(AdresRepository adresRepository) {
+        super(adresRepository);
     }
 }
