+++ microservices-gateway/pom.xml
@@ -24,10 +24,12 @@
             <artifactId>spring-cloud-starter-netflix-eureka-server</artifactId>
         </dependency>
 
+        <!--
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-netflix-zuul</artifactId>
         </dependency>
+-->
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
