+++ microservices-gateway/src/main/resources/application.properties
@@ -9,7 +9,7 @@ eureka.client.service-url.defaultZone=http://localhost:8080/eureka
 #timeout
 eureka.server.wait-time-in-ms-when-sync-empty=0
 #All url come with prefix/api will interpret
-zuul.prefix=/api
-
-zuul.routes.middleware.path=/demo/**
-zuul.routes.middleware.url=http://localhost:8081/
+#zuul.prefix=/api
+#
+#zuul.routes.middleware.path=/demo/**
+#zuul.routes.middleware.url=http://localhost:8081/
